refactor(email): move body_parsed tests to test.rs, log panic via inspect_err
